requirements.txt
SpeechRecognition
groq
pyaudio
python-dotenv